Added PrEP incrementing

+++ transmission_model/src/GeometricDistribution.cpp
@@ -11,7 +11,7 @@
 
 namespace TransModel {
 
-GeometricDistribution::GeometricDistribution(double p, int increment) : dist(p), increment_{increment} {}
+GeometricDistribution::GeometricDistribution(double p, double increment) : dist(p), increment_{increment} {}
 
 GeometricDistribution::~GeometricDistribution() {
 }
